# docs/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "سامانه مدیریت دانش و نظام مسائل (دانا) - API Specification",
    "description": "مستندات رسمی و دقیق Swagger / OpenAPI 3.0 برای وب‌سرویس‌های نظام مسائل، گردش کار، گزارشات، پیوست‌ها و خروجی‌های سامانه دانا",
    "version": "5.0.0"
  },
  "servers": [
    {
      "url": "/api",
      "description": "سرور محلی و عملیاتی دانا"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Issue": {
        "type": "object",
        "required": ["title", "domainNodeId"],
        "properties": {
          "id": { "type": "integer", "readOnly": true },
          "title": { "type": "string", "example": "طراحی و توسعه ماژول رمزنگاری کوانتومی" },
          "domainNodeId": { "type": "integer", "description": "شناسه گره دانشی در درختواره" },
          "periodId": { "type": "integer", "nullable": true, "description": "شناسه دوره زمانی" },
          "researchItemId": { "type": "integer", "nullable": true, "description": "شناسه آیتم پژوهشی مرتبط" },
          "sourceIssueId": { "type": "integer", "nullable": true, "description": "شناسه مسئله مرجع در صورت انتقال بین‌دوره‌ای" },
          "gapId": { "type": "integer", "nullable": true, "description": "شناسه گپ دانشی مرتبط" },
          "solutionDirection": { "type": "string", "nullable": true },
          "responsibleUnit": { "type": "string", "nullable": true },
          "confidentialityLevel": { "type": "string", "enum": ["عمومی", "محرمانه", "خیلی محرمانه", "سری", "به‌کلی سری"], "default": "عمومی" },
          "actionPriority": { "type": "string", "enum": ["خیلی زیاد", "زیاد", "متوسط", "کم"], "default": "متوسط" },
          "approvalDate": { "type": "string", "nullable": true },
          "knowledgeType": { "type": "string", "nullable": true },
          "projectLevel": { "type": "string", "nullable": true },
          "approvalAuthority": { "type": "string", "nullable": true },
          "researchProjectType": { "type": "string", "nullable": true },
          "knowledgeProjectType": { "type": "string", "nullable": true },
          "events": { "type": "string", "nullable": true },
          "scientificDiplomacy": { "type": "string", "nullable": true },
          "collaborators": { "type": "string", "nullable": true },
          "referenceDocument": { "type": "string", "nullable": true },
          "actionsTaken": { "type": "string", "nullable": true },
          "bottlenecks": { "type": "string", "nullable": true },
          "orders": { "type": "string", "nullable": true },
          "category": { "type": "string", "nullable": true },
          "requiredBudget": { "type": "number", "default": 0 },
          "approvedBudget": { "type": "number", "default": 0 },
          "assignedBudget": { "type": "number", "default": 0 },
          "expectedMonths": { "type": "integer", "default": 0 },
          "completionPercent": { "type": "integer", "minimum": 0, "maximum": 100, "default": 0 },
          "status": { "type": "string", "enum": ["pending", "in_progress", "completed", "canceled", "on_hold"], "default": "pending" },
          "needStatement": { "type": "object", "description": "اطلاعات تب بیان نیاز" },
          "contract": { "type": "object", "description": "اطلاعات تب قرارداد" },
          "executiveContract": { "type": "object", "description": "اطلاعات تب قرارداد اجرایی" },
          "stage20": { "type": "object", "description": "اطلاعات مقطع ۲۰٪" },
          "stage50": { "type": "object", "description": "اطلاعات مقطع ۵۰٪" },
          "stage100": { "type": "object", "description": "اطلاعات مقطع ۱۰۰٪" },
          "application": { "type": "object", "description": "اطلاعات تب کاربست" },
          "macroProject": { "type": "object", "description": "اطلاعات تب کلان‌پروژه" },
          "collaborationNetwork": { "type": "object", "description": "شبکه همکاران" },
          "issueResolutionTeam": { "type": "array", "items": { "type": "object" }, "description": "تیم حل مسئله" },
          "templateIds": { "type": "array", "items": { "type": "integer" } },
          "metadata": { "type": "object" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
...
    }
  },
  "security": [{ "bearerAuth": [] }],
  "paths": {
    "/issues": {
      "get": {
        "summary": "دریافت فهرست مسائل با صفحه‌بندی و کنترل دسترسی سازمانی",
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 20 } },
          { "name": "periodId", "in": "query", "schema": { "type": "integer" } },
          { "name": "domain", "in": "query", "schema": { "type": "string" } },
          { "name": "status", "in": "query", "schema": { "type": "string" } },
          { "name": "priority", "in": "query", "schema": { "type": "string" } },
          { "name": "category", "in": "query", "schema": { "type": "string" } },
          { "name": "search", "in": "query", "schema": { "type": "string" } },
          { "name": "advancedFilter", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "لیست مسائل همراه با اطلاعات صفحه‌بندی" }
        }
      },
      "post": {
        "summary": "ثبت مسئله جدید",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Issue" } } }
        },
        "responses": {
          "201": { "description": "مسئله با موفقیت ایجاد شد" },
          "400": { "description": "خطای اعتبارسنجی فیلدها" },
          "403": { "description": "عدم دسترسی به حوزه دانشی مربوطه" }
        }
      }
    },
...
  }
}